Retrieve Historical Time Series Data

Fetches historical data for one or multiple RICs from the API.

Usage

get_timeseries(
  rics,
  fields = "*",
  start_date = NULL,
  end_date = NULL,
  interval = "daily",
  raw_output = FALSE,
  calendar = NULL,
  count = NULL,
  corax = NULL,
  debug = FALSE,
  verbose = FALSE
)

Arguments

rics

A character string or list of character strings specifying the RICs to retrieve data for.

fields

A character string or list of strings specifying the fields to retrieve. Available fields: `c("TIMESTAMP", "VALUE", "VOLUME", "HIGH", "LOW", "OPEN", "CLOSE", "COUNT")`. Default: `"*"` (all fields).

start_date

A character string specifying the start date in format `" Default: 100 days before the current date.

end_date

A character string specifying the end date in format `" Default: current date.

interval

A character string specifying the data interval. Possible values: `"tick"`, `"minute"`, `"hour"`, `"daily"`, `"weekly"`, `"monthly"`, `"quarterly"`, `"yearly"`. Default: `"daily"`.

raw_output

A logical value. If `TRUE`, returns data in raw JSON format. If `FALSE`, returns a data frame. Default: `FALSE`. This parameter is ignored if `raw_output = TRUE`. Default: `FALSE`.

calendar

A character string specifying the calendar type. Possible values: `"native"`, `"tradingdays"`, `"calendardays"`.

count

An integer specifying the maximum number of data points to retrieve.

corax

A character string specifying the correction method. Possible values: `"adjusted"`, `"unadjusted"`.

debug

A logical value. If `TRUE`, prints debug messages, including JSON request and response. Default: `FALSE`.

verbose

A logical value. If `TRUE`, prints messages. Default: `FALSE`.

Value

A data frame with historical data or a JSON response if `raw_output = TRUE`.